Expose auto-capture-workflow via pytest plugin to downstream proects via plug&play
