+++ docs/README.md
@@ -13,6 +13,14 @@ Welcome to the Chroma MCP Server documentation. This guide provides comprehensiv
 - [Docker](#docker)
 - [Logging](#logging)
 - [Working Memory](#working-memory)
+- [IDE & Tool Integration Guides](integration/README.md)
+- [Usage Guides](usage/README.md)
+- [Client and Developer Scripts](scripts/README.md)
+- [Thinking Tools & Utilities](thinking_tools/README.md)
+- [Logging Documentation](logging/README.md)
+- [Automation Documentation](automation/README.md)
+- [Project Rules & Guidelines](rules/README.md)
+- [Refactoring Plans](refactoring/README.md)
 - [Enhanced Context Capture](usage/enhanced_context_capture.md)
 - [Semantic Code Chunking](usage/semantic_chunking.md)
 - [Tool Usage Format Specification](usage/tool_usage_format.md)
