+++ docs/automation/README.md
@@ -1,0 +1,8 @@
+# Automation Documentation
+
+This section details various automation aspects of the `chroma-mcp-server` project, focusing on how to automate common development tasks and workflows.
+
+## Available Guides
+
+- **[Git Hooks](./git_hooks.md)**
+  - Explains the setup and usage of Git hooks, such as `post-commit` hooks for incremental codebase indexing and test workflow automation.
