+++ docs/logging/README.md
@@ -1,0 +1,8 @@
+# Logging Documentation
+
+This section covers the logging mechanisms within `chroma-mcp-server`.
+
+## Available Guides
+
+- **[Server Logging](./server_logging.md)**
+  - Explains how server-side logging is handled, including log redirection in stdio mode, log levels, and log file management like rotation.
