+++ docs/rules/README.md
@@ -1,0 +1,22 @@
+# Project Rules and Guidelines
+
+This section contains documentation on various rules, guidelines, and best practices adopted by the `chroma-mcp-server` project. These are intended to ensure consistency, quality, and smooth collaboration.
+
+The rules are mentioned by the `.cursorrules`, `.windsurfrules` and `.github/copilot-instructions.md` files in this project for use with Cursor, Windsurf and GitHub Copilot respectively.
+
+## Available Guides
+
+- **[Testing and Build Guide](./testing-and-build-guide.md)**
+  - Provides comprehensive instructions on how to build the project, run tests (including coverage and specific versions), and manage development environments using Hatch.
+
+- **[Automated Chat Logging Rule (`auto_log_chat`)](./auto_log_chat.md)**
+  - Describes the rule that enables automatic logging of AI chat summaries to ChromaDB, enhancing context capture for RAG.
+
+- **[Memory Integration Rules](./memory-integration-rules.md)**
+  - Outlines how the AI should leverage previous thoughts and conceptual discussions, particularly using sequential thinking tools and querying similar thoughts.
+
+- **[Markdown Linting Rules](./markdown-rules.md)**
+  - Details the linting rules for writing consistent and maintainable Markdown files within the project.
+
+- **[Python and GitHub Development Rules](./python-github-rules.md)**
+  - Covers development guidelines specific to Python code and GitHub practices, such as commit messages, pull requests, and branch management.
